+++ shardingsphere-jdbc/shardingsphere-jdbc-spring/shardingsphere-jdbc-core-spring/shardingsphere-jdbc-core-spring-namespace/src/main/java/org/apache/shardingsphere/driver/spring/namespace/tag/TransactionTypeScannerBeanDefinitionTag.java
@@ -15,22 +15,16 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.driver.spring.namespace.constants;
+package org.apache.shardingsphere.driver.spring.namespace.tag;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 
 /**
- * Sharding rule parser tag constants.
+ * Transaction type scanner bean definition tag.
  */
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
-public final class DataSourceBeanDefinitionParserTag {
+public final class TransactionTypeScannerBeanDefinitionTag {
     
-    public static final String ROOT_TAG = "data-source";
-    
-    public static final String DATA_SOURCE_NAMES_TAG = "data-source-names";
-    
-    public static final String RULE_REFS_TAG = "rule-refs";
-    
-    public static final String PROPS_TAG = "props";
+    public static final String ROOT_TAG = "tx-type-annotation-driven";
 }
